Create README.md
# MRA_code
This is the source code of Mud Ring Algorithm (MRA) written in MATLAB.
Mud Ring Algorithm (MRA) is a novel optimization algorithm that mimics the mud ring feeding behaviour of bottlenose dolphins in the Atlantic coast of Florida.
The structure of MRA is simple and easy to implement. 
You can download the main paper [here](https://doi.org/10.1109/ACCESS.2022.3173401).